Provide a tthave() instead of a global have variable
This should make it easier to keep TTY input status local only to tty.c,
easing the transition to moving away from the hack to get timeouts that
joe is currently using:

	set_nonblocking(stdin)
	have = (read(havec) == 1)
	set_blocking(stdin)

To something else, such as poll(), select(), or similar. This should clean
up the mess that is the use of signals and flags for timeouts as well.